research/test.py
from google import genai
from google.genai import types
import os

# os.environ[""]

client = genai.Client(api_key="AIzaSyBWZs2IlMeGYal2aHM8cWR5h9A4p-kcBPs")





response = client.models.generate_content(
    model="gemini-3-pro-image-preview",
    contents="Generate an infographic of the current weather in Tokyo.",
    config=types.GenerateContentConfig(
        tools=[{"google_search": {}}],
        image_config=types.ImageConfig(
            aspect_ratio="16:9",
            image_size="4K"
        )
    )
)


image_parts = [part for part in response.parts if part.inline_data]

if image_parts:
    image = image_parts[0].as_image()
    image.save('weather_tokyo.png')
    image.show()