pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "boarding-mcp-server"
version = "0.1.0"
description = "MCP Server for Onboarding Management"
readme = "README.md"
authors = [
    { name = "Girish", email = "girish12n@gmail.com" }
]
requires-python = ">=3.8"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "sqlalchemy",
    "alembic",
    "psycopg2-binary",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "httpx",
    "fastmcp",
    "mcp",
    "google-genai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "isort",
]

test = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[project.scripts]
boarding-server = "app.main:main"

[tool.setuptools.packages.find]
exclude = ["tests*", "*.tests", "*.tests.*"]
